Remove window.insights.chrome references.
