+++ src/auth/index.ts
@@ -11,6 +11,7 @@ import consts, { defaultAuthOptions as defaultOptions } from '../utils/consts';
 import { ACCOUNT_REQUEST_TIMEOUT, ACTIVE_REMOTE_REQUEST, CROSS_ACCESS_ACCOUNT_NUMBER, CROSS_ACCESS_ORG_ID } from '../utils/consts';
 import qe from '../utils/iqeEnablement';
 import { ChromeModule } from '../@types/types';
+import { createFetchPermissionsWatcher } from './fetchPermissions';
 
 export type LibJWT = {
   getOfflineToken: () => Promise<AxiosResponse<any>>;
@@ -61,6 +62,19 @@ export const createGetUser = (libjwt: LibJWT): (() => Promise<ChromeUser | undef
     libjwt.initPromise.then(libjwt.jwt.getUserInfo).catch(() => {
       libjwt.jwt.logoutAllTabs();
     });
+};
+
+export const createGetUserPermissions = (libJwt: LibJWT, getUser: () => Promise<void | ChromeUser>) => {
+  const fetchPermissions = createFetchPermissionsWatcher(getUser);
+  return async (app = '', bypassCache?: boolean) => {
+    if (isITLessEnv) {
+      const cogToken = await getTokenWithAuthorizationCode();
+      return fetchPermissions(cogToken || '', app, bypassCache);
+    } else {
+      await getUser();
+      return fetchPermissions(libJwt.jwt.getEncodedToken() || '', app, bypassCache);
+    }
+  };
 };
 
 export default ({ ssoUrl }: { ssoUrl?: string }): LibJWT => {
