+++ src/utils/consts.ts
@@ -1,12 +1,5 @@
-import instance from '@redhat-cloud-services/frontend-components-utilities/interceptors';
-import get from 'lodash/get';
-import isEmpty from 'lodash/isEmpty';
-import cookie from 'js-cookie';
-import { getFeatureFlagsError, unleashClient } from '../components/FeatureFlags/FeatureFlagsProvider';
-import { ITLess, isBeta, isProd } from './common';
-import { AxiosRequestConfig } from 'axios';
+import { ITLess } from './common';
 import { AppNavigationCB, ChromeAuthOptions, GenericCB, NavDOMEvent } from '../@types/types';
-import { VisibilityFunctions } from '@redhat-cloud-services/types';
 import { Listener } from '@redhat-cloud-services/frontend-components-utilities/MiddlewareListener';
 import { APP_NAV_CLICK, GLOBAL_FILTER_UPDATE } from '../redux/action-types';
 
@@ -31,108 +24,6 @@ export const ACCESS_KEY = 'access_key';
 const obj = {
   noAuthParam,
   offlineToken,
-};
-
-const matcherMapper = {
-  isEmpty,
-  isNotEmpty: (value: any) => !isEmpty(value),
-};
-
-const matchValue = (value: any, matcher?: keyof typeof matcherMapper) => {
-  const match = matcherMapper[matcher!];
-  return typeof match === 'function' ? match(value) : value;
-};
-
-/**
- * Check if is permitted to see navigation link
- * @param {array} permissions array checked user permissions
- * @param {every|some} require type of permissions requirement
- * @returns {boolean}
- */
-const checkPermissions = async (permissions: string[] = [], require: 'every' | 'some' = 'every') => {
-  const userPermissions = await window.insights.chrome.getUserPermissions();
-  return userPermissions && permissions[require]((item) => userPermissions.find(({ permission }) => permission === item));
-};
-
-export const visibilityFunctions: VisibilityFunctions = {
-  isOrgAdmin: async () => {
-    const data = await window.insights.chrome.auth.getUser();
-    try {
-      return !!data?.identity.user?.is_org_admin;
-    } catch {
-      return false;
-    }
-  },
-  isActive: async () => {
-    const data = await window.insights.chrome.auth.getUser();
-    try {
-      return !!data?.identity.user?.is_active;
-    } catch {
-      return false;
-    }
-  },
-  isInternal: async () => {
-    const data = await window.insights.chrome.auth.getUser();
-    try {
-      return !!data?.identity.user?.is_internal;
-    } catch {
-      return false;
-    }
-  },
-  isEntitled: async (appName?: string) => {
-    const data = await window.insights.chrome.auth.getUser();
-    const { entitlements } = data || { entitlements: {} };
-    return data?.entitlements && appName
-      ? Boolean(entitlements[appName] && entitlements[appName].is_entitled)
-      : // eslint-disable-next-line camelcase
-        Object.entries(entitlements || {}).reduce((acc, [key, { is_entitled }]) => ({ ...acc, [key]: is_entitled }), {});
-  },
-  isProd: () => isProd(),
-  isBeta: () => isBeta(),
-  isHidden: () => true, // FIXME: Why always true?
-  withEmail: async (...toHave: string[]) => {
-    const data = await window.insights.chrome.auth.getUser();
-    const {
-      identity: { user },
-    } = data || { identity: {} };
-    return toHave?.some((item) => user?.email?.includes(item));
-  },
-  loosePermissions: (permissions: string[]) => checkPermissions(permissions, 'some'),
-  hasPermissions: checkPermissions,
-  hasLocalStorage: (key: string, value: unknown) => localStorage.get(key) === value,
-  hasCookie: (cookieKey: string, cookieValue: string) => cookie.get(cookieKey) === cookieValue,
-  apiRequest: async ({
-    url,
-    method = 'GET',
-    accessor,
-    matcher,
-    ...options
-  }: Omit<AxiosRequestConfig, 'adapter'> & { accessor?: string; matcher?: keyof typeof matcherMapper }) => {
-    const data = await window.insights.chrome.auth.getUser();
-
-    const token = await window.insights.chrome.auth.getToken();
-    // this will log a bunch of 403s if the account number isn't present
-    if (data?.identity.org_id) {
-      return instance({
-        url,
-        method,
-        ...options,
-        headers: {
-          Authorization: `Bearer ${token}`,
-          ...options.headers,
-        },
-      })
-        .then((response) => matchValue(accessor ? get(response || {}, accessor) : response, matcher))
-        .catch(() => {
-          console.log('Unable to retrieve visibility result', { visibilityMethod: 'apiRequest', method, url });
-          return false;
-        });
-    } else {
-      console.log('Unable to call API, no account number');
-      return false;
-    }
-  },
-  featureFlag: (flagName: string, expectedValue: boolean) => getFeatureFlagsError() !== true && unleashClient?.isEnabled(flagName) === expectedValue,
 };
 
 export const isVisible = (limitedApps?: string[], app?: string, visibility?: Record<string, unknown>) => {
