+++ src/utils/internalChromeContext.ts
@@ -1,0 +1,6 @@
+import { ChromeAPI } from '@redhat-cloud-services/types';
+import { createContext } from 'react';
+
+const InternalChromeContext = createContext<ChromeAPI>({} as ChromeAPI);
+
+export default InternalChromeContext;
